split functions to rendering and sidebar, use index.js as single export file
